Restructured the project architecture, layout.
