+++ Packagebird-Client/DOC.md
@@ -1,0 +1,2 @@
+# The Client component of the Packagebird solution
+Contains all items for the 'frontend' of the application.
