+++ Packagebird-Server/DOC.md
@@ -1,0 +1,2 @@
+# The Server component of Packagebird
+Contains all items for the 'backend' of the application.
